redirect to student director
