+++ src/app/components/Students.js
@@ -1,5 +1,5 @@
 import React from 'react';
 
-const Students = () => <h1>All students</h1>;
+const Student = () => <h1>Students</h1>;
 
-export default Students;
+export default Student;
